pyproject.toml
[project]
name = "pywilbox"
version = "0.1.0"
description = "Control interface for the Wilbox hardware"
authors = [{ name="Matthias Brennwald", email="mbrennwa@gmail.com" }]
readme = "README.md"
dependencies = [
    "pyrelayctl @ git+https://github.com/xypron/pyrelayctl.git",
    "pyserial"
]
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
